refactor: QtCore Executor out of hotkey submodule
It provides a thread-safe way to run functions
inside the Qt event loop from a different thread.
