+++ setup.cfg
@@ -32,8 +32,13 @@ good-names =
   asyncSetUp,  # Used by unittest.
   do_EOF,  # Used by cmd.
   setUp,  # Used by unittest.
+  # PySide2
+  QtCore,
+  test_QtCore,
   # PySide2.QtCore
   QCoreApplication,
+  # PySide2.QtCore.QObject
+  customEvent,
   # PySide2.QtWidgets.QWidget
   closeEvent,
   hideEvent,
